refactor: Improve gift target search
It will now search for targets using any part of their names, not only the first few letters. It also will show a list of matches if more than one is found. Clicking on a name on the list will copy and paste to chat bar the same command the sender used, with the target argument replaced.