refactor: improve server version checker
As of MC 1.20.5 PaperMC is following the Mojang mappings instead of the Spigot mappings. One effect of this change is the classes outside the API are no longer versioned. The old method for retrieving the NMS version would return "craftbukkit" due to this, which made it confusing while debugging. The new method will return "unknown" instead.