feat: make most plugin messages translatable
Some messages that appear in console during the startup are still hardcoded, however messages visible to the end users can be customized and translated in different languages now. 2 languages are available and embedded into the plugin - en_US and sas_SY, which is also English but gives the plugin an attitude. Additionally, support for RGB color codes has been added. At the moment there isn't a way to copy an embedded file to disk for customization, but it is planned.